feedback(kaex): Expose data and improve default error msg for env not found error
