+++ pkg/okctl/okctl.go
@@ -51,9 +51,6 @@ import (
 	"github.com/oslokommune/okctl/pkg/storage"
 )
 
-// ErrorEnvironmentNotFound indicates that the requested environment isn't available
-var ErrorEnvironmentNotFound = errors.New("environment not found")
-
 // Okctl stores all state required for invoking commands
 type Okctl struct {
 	*config.Config
@@ -70,7 +67,10 @@ type Okctl struct {
 // set previously
 func (o *Okctl) InitialiseWithOnlyEnv(env string) error {
 	if !o.RepoState.HasEnvironment(env) {
-		return fmt.Errorf("%w: %s", ErrorEnvironmentNotFound, env)
+		return ErrorEnvironmentNotFound{
+			TargetEnvironment:     env,
+			AvailableEnvironments: getEnvironments(o.RepoState.Clusters),
+		}
 	}
 
 	repoDir, err := o.GetRepoDir()
